Move the Functor, Foldable, HEqF, and HShowF instances into the Pattern module.
