+++ src/Derivative/Parser.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE DeriveFunctor, FlexibleInstances, GADTs, MultiParamTypeClasses, RankNTypes, ScopedTypeVariables #-}
+{-# LANGUAGE DeriveFunctor, FlexibleInstances, GADTs, RankNTypes, ScopedTypeVariables #-}
 module Derivative.Parser
 ( cat
 , commaSep
@@ -208,29 +208,6 @@ newtype K a b = K { getK :: a }
 
 -- Instances
 
-instance HFunctor (PatternF t) where
-  hfmap f p = case p of
-    Cat a b -> Cat (f a) (f b)
-    Alt a b -> Alt (f a) (f b)
-    Rep a -> Rep (f a)
-    Map g p -> Map g (f p)
-    Bnd p g -> Bnd (f p) (f . g)
-    Sat p -> Sat p
-    Ret as -> Ret as
-    Nul -> Nul
-    Lab p s -> Lab (f p) s
-    Del a -> Del (f a)
-
-instance (Alternative a, Monad a) => HFoldable (PatternF t) a where
-  hfoldMap f p = case p of
-    Cat a b -> f ((,) <$> a <*> b)
-    Alt a b -> f (a <|> b)
-    Map g p -> f (g <$> p)
-    Bnd p g -> f (p >>= g)
-    Lab p _ -> f p
-    Del a -> f a
-    _ -> empty
-
 instance Functor (Rec (PatternF t) v) where
   fmap f = compact' . rec . Map f
 
@@ -264,35 +241,6 @@ instance Monad (Rec (PatternF t) v) where
 instance Monad (Graph (PatternF t)) where
   return = pure
   Graph p >>= f = Graph (p >>= unGraph . f)
-
-instance HEqF (PatternF t)
-  where heqF eq a b = case (a, b) of
-          (Cat a1 b1, Cat a2 b2) -> eq a1 a2 && eq b1 b2
-          (Alt a1 b1, Alt a2 b2) -> eq a1 a2 && eq b1 b2
-          -- (Map f1 p1, Map f2 p2) -> eq p1 p2
-          -- (Bnd p1 f1, Bnd p2 f2) -> eq p1 p2
-          (Sat p1, Sat p2) -> p1 == p2
-          (Ret r1, Ret r2) -> length r1 == length r2
-          (Nul, Nul) -> True
-          (Lab p1 s1, Lab p2 s2) -> s1 == s2 && eq p1 p2
-          _ -> False
-
-instance Show t => HShowF (PatternF t)
-  where hshowsPrecF showsPrec n p = case p of
-          Cat a b -> showParen (n > 4) $ showsPrec 4 a . showString " `cat` " . showsPrec 5 b
-          Alt a b -> showParen (n > 3) $ showsPrec 3 a . showString " <|> " . showsPrec 4 b
-          Rep a -> showParen (n >= 10) $ showString "many " . showsPrec 10 a
-          Map _ p -> showParen (n > 4) $ showString "f <$> " . showsPrec 5 p
-          Bnd p _ -> showParen (n > 1) $ showsPrec 1 p . showString " >>= f"
-          Sat (Equal c) -> showParen (n >= 10) $ showString "char " . shows c
-          Sat (Category c) -> showParen (n >= 10) $ showString "category " . shows c
-          Sat (Constant _) -> showString "anyToken"
-          Ret [_] -> showParen (n >= 10) $ showString "pure t"
-          Ret t -> showString "ret [" . showIndices (length t) . showString "]"
-          Nul -> showString "empty"
-          Lab p s -> showParen (n > 2) $ showsPrec 3 p . showString " `label` " . shows s
-          Del a -> showParen (n >= 10) $ showString "delta " . showsPrec 10 a
-          where showIndices n = foldr (.) id ((showChar 't' .) . shows <$> take n (iterate succ (0 :: Integer)))
 
 instance Monoid a => Applicative (K a)
   where pure = const (K mempty)
